+++ app/Http/Controllers/Api/AsyncUploadRedactorMediaController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Thinktomorrow\Chief\App\Http\Controllers\Back\Media;
+namespace Thinktomorrow\Chief\App\Http\Controllers\Api;
 
 use Illuminate\Http\Request;
 use Thinktomorrow\Chief\Media\MediaType;
@@ -9,7 +9,7 @@ use Thinktomorrow\AssetLibrary\Application\AddAsset;
 use Thinktomorrow\AssetLibrary\Application\AssetUploader;
 use Thinktomorrow\Chief\App\Http\Controllers\Controller;
 
-class UploadManagersMediaController extends Controller
+class AsyncUploadRedactorMediaController extends Controller
 {
     /** @var Managers */
     private $managers;
@@ -30,12 +30,12 @@ class UploadManagersMediaController extends Controller
      * @throws \Spatie\MediaLibrary\Exceptions\FileCannotBeAdded
      * @throws \Thinktomorrow\AssetLibrary\Exceptions\AssetUploadException
      */
-    public function store(string $key, $id, Request $request)
+    public function upload(string $key, $id, Request $request)
     {
-        $uploads = $request->file('file');
-        $model   = $this->managers->findByKey($key, $id)->existingModel();
+        $files = $request->input('files', []);
+        $model = $this->managers->findByKey($key, $id)->existingModel();
 
-        if (empty($uploads)) {
+        if (!is_array($files) || empty($files)) {
             return response()->json([
                 'error'    => true,
                 'messages' => 'Geen afbeelding opgeladen.',
@@ -44,12 +44,16 @@ class UploadManagersMediaController extends Controller
 
         $responseContent = [];
 
-        foreach ($uploads as $upload) {
-            if (! $asset = AssetUploader::upload($upload)) {
-                return response()->json([
+        foreach($files as $filePayload) {
+            $base64EncodedFile = $filePayload['data'];
+            $filename = $filePayload['filename'];
+
+            if (! $asset = AssetUploader::uploadFromBase64($base64EncodedFile, $filename)) {
+                $responseContent['file-' . rand(1-999)] = [
                     'error'    => true,
-                    'messages' => 'Afbeelding kan niet worden opgeladen.',
-                ], 200);
+                    'messages' => 'Afbeelding ['.$filename.'] kan niet worden opgeladen.',
+                ];
+                continue;
             }
 
             app(AddAsset::class)->add($model, $asset, MediaType::CONTENT, $request->input('locale', app()->getLocale()));