+++ resources/assets/js/native.js
@@ -3,4 +3,5 @@
  * loaded after the Vue instantiation
  */
 
+require('./utilities/chiefRedactorImageUpload');
 require('./utilities/form-submit');