+++ routes/chief-admin-routes.php
@@ -57,7 +57,7 @@ Route::get('menuitem/{id}/edit', 'Thinktomorrow\Chief\App\Http\Controllers\Back\
 * EDITOR API & MEDIA MANAGEMENT (used by editor)
 * -----------------------------------------------------------------
 */
-Route::post('managers/{key}/{id}/media', 'Thinktomorrow\Chief\App\Http\Controllers\Back\Media\UploadManagersMediaController@store')->name('chief.back.managers.media.upload');
+Route::post('managers/{key}/{id}/media', 'Thinktomorrow\Chief\App\Http\Controllers\Api\AsyncUploadRedactorMediaController@upload')->name('chief.back.managers.media.upload');
 Route::post('api/assets/images/upload', 'Thinktomorrow\Chief\App\Http\Controllers\Api\AsyncUploadSlimMediaController@upload')->name('chief.api.images.upload');
 Route::post('api/assets/files/upload', 'Thinktomorrow\Chief\App\Http\Controllers\Api\AsyncUploadFileMediaController@upload')->name('chief.api.files.upload');
 Route::get('api/internal-links', 'Thinktomorrow\Chief\App\Http\Controllers\Api\InternalLinksController@index')->name('chief.api.internal-links');